pyproject.toml
[project]
name = "codex-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.4.0",
    "psycopg[binary]>=3.2.0",
    "gunicorn>=22.0.0",
    "django-environ>=0.12.0",
    "drf-yasg>=1.21.11",
    "pytest-django>=4.11.1",
]
